% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_counts}
\alias{add_counts}
\alias{add_counts.character}
\alias{add_counts.factor}
\title{Add counts in parentheses to a factor or character vector}
\usage{
add_counts(x)

\method{add_counts}{character}(x)

\method{add_counts}{factor}(x)
}
\arguments{
\item{x}{A factor or character vector}
}
\description{
Add counts in parentheses to a factor or character vector
}
