% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_country}
\alias{convert_country}
\title{Wrap the countrycode package to convert back and forth between country
code and country name}
\usage{
convert_country(x)
}
\arguments{
\item{x}{Vector of either 2-letter country code or country names}
}
\description{
If given a vector of country codes, turn into names. If given a vector
of names, turn back into country codes. This is a very simple utility
since I find myself using this code too often
}

