% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{decode_base_62}
\alias{decode_base_62}
\title{Decode a string in base 62 into a number}
\usage{
decode_base_62(s)
}
\arguments{
\item{s}{a character vector in base 62 encoding}
}
\description{
Note that this returns a float vector, not an integer, due to
the (unfortunate) 32-bit limit of integers in R.
}
\examples{

strings <- c("wojWiVquxXy", "x1VMNGRqI3m")

decode_base_62(strings)

}
