% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{expand_grid}
\alias{expand_grid}
\title{Wrapper around expand.grid more suited to factorial experiments}
\usage{
expand_grid(..., stringsAsFactors = FALSE, group_by_all = FALSE)
}
\arguments{
\item{...}{Options passed on to \code{\link{expand.grid}}}

\item{stringsAsFactors}{Whether to treat strings as factors, by default
off (thankfully)}

\item{group_by_all}{whether to group}
}
\description{
Wrapper around \code{\link{expand.grid}} that returns a tbl_df instead of a
data.frame, that allows grouping by all columns as an option, and that
treats stringsAsFactors as FALSE by default.
}

