% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-strata.R
\name{extract_strata}
\alias{extract_strata}
\title{Extract strata from a tidied survival object}
\usage{
extract_strata(tbl, column, ...)
}
\arguments{
\item{tbl}{A table}

\item{column}{Column to extract. Should be of the form \code{col1=val1, col2=val2, ...}}

\item{...}{Extra arguments passed on to \link[tidyr]{extract}, such as \code{remove} or
\code{convert}.df}
}
\description{
Extract strata from a tidied survival object
}
\examples{

library(survival)
library(broom)

surv_fit <- survfit(Surv(stop, status) ~ sex + pcdx,
                    data = mgus1, subset = (start == 0))

surv_fit \%>\%
  tidy() \%>\%
  extract_strata(strata)

}
