% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{\%<*>\%}
\alias{\%<*>\%}
\title{Multiplies matrices after matching the rows of one to the columns of the other}
\usage{
m1 \%<*>\% m2
}
\arguments{
\item{m1}{First matrix}

\item{m2}{Second matrix}
}
\value{
A matrix with \code{nrow(m1)} rows and \code{ncol(m2)} columns.
}
\description{
This is a custom infix operator: %<*>% multiplies matrices, but first matches the
column names of the first with the rows of the second. This works only if the first
matrix has column names and the second has row names, so they can be matched.
}
