% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{json_vec_to_df}
\alias{json_vec_to_df}
\title{Turn a character vector of JSON dictionaries into a data frame}
\usage{
json_vec_to_df(j)
}
\arguments{
\item{j}{Character vector containing JSON dictionaries}
}
\description{
Sometimes (e.g. in a database column) one has a vector of individual
JSON dictionaries with the same or similar keys.
This is a simple utility to parse them all at
once into a data frame. It's substantially more efficient than
individually parsing them and then recombining them.
}
\examples{

input <- c('{"col1": 3, "col2": 10}', '{"col1": 5}',
           NA, '{"col1": 50, "col2": 100}')

json_vec_to_df(input)

}
