% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{melt_multiple}
\alias{melt_multiple}
\title{Melt multiple matrices of the same dimension into multiple columns}
\usage{
melt_multiple(..., varnames = NULL, tbldf = TRUE)
}
\arguments{
\item{...}{Matrices, given as named arguments}

\item{varnames}{Variable names to use in molten data.frame}

\item{tbldf}{Whether to return a tbl_df instead of a data.frame}
}
\description{
Melt multiple matrices of the same dimension into multiple columns
}

