% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{relibrary}
\alias{relibrary}
\title{Detach and then reload a package based on a character string of its name}
\usage{
relibrary(package_name)
}
\arguments{
\item{package_name}{A character string (not a bare name, yet)}
}
\description{
If the package is not already loaded, it still loads it
}
