% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{reorder_col}
\alias{reorder_col}
\title{Reorder a column of a tbl based on another}
\usage{
reorder_col(.data, col, value, ...)
}
\arguments{
\item{.data}{A tbl}

\item{col}{Column to use for reordering}

\item{value}{Value to reorder by}

\item{...}{Extra arguments passed on to reorder, such as \code{function}}
}
\description{
Reorder a column of a tbl based on another
}

