% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{rotate_x_labels}
\alias{rotate_x_labels}
\title{Rotate the labels on the x-axis of a graph}
\usage{
rotate_x_labels(...)
}
\arguments{
\item{...}{Extra arguments passed on to element_text}
}
\description{
Rotate the labels on the x-axis of a graph
}
