% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{scale_percent}
\alias{scale_percent}
\alias{scale_x_percent}
\alias{scale_xy_percent}
\alias{scale_y_percent}
\title{Percentage scales}
\usage{
scale_x_percent(...)

scale_y_percent(...)

scale_xy_percent(...)
}
\description{
Format the labels on the x axis, y axis, or both
}

