% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{scale_xy_log10}
\alias{scale_xy_log10}
\title{Log scale for both x & y axes}
\usage{
scale_xy_log10(...)
}
\arguments{
\item{...}{Arguments passed on to the scales}
}
\description{
Transform both the x and y axis for the common case of a log-log plot
}

