% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sparse_cast_}
\alias{sparse_cast_}
\title{Standard-evaluation version of sparse_cast}
\usage{
sparse_cast_(data, row_col, column_col, value_col = NULL)
}
\arguments{
\item{data}{A tbl}

\item{row_col}{String version of column to use as row names}

\item{column_col}{String version of column to use as column names}

\item{value_col}{String version of column to use as sparse matrix values,
or a numeric value to use}
}
\description{
Standard-evaluation version of sparse_cast
}

