% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{split_unnest}
\alias{split_unnest}
\title{Split a string column and then unnest on that split column}
\usage{
split_unnest(data, col, sep = "[^[:alnum:]]+")
}
\description{
Split a string column and then unnest on that split column
}

