% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{theme_blank}
\alias{theme_blank}
\title{Blank theme: used to come with ggplot2}
\usage{
theme_blank(...)
}
\arguments{
\item{...}{Arguments passed to theme_bw}
}
\description{
Blank theme: used to come with ggplot2
}
