% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tidy_cor}
\alias{tidy_cor}
\title{Perform a correlation and melt the result into a table}
\usage{
tidy_cor(m, use = "everything", method = "pearson", varnames = c("Var1",
  "Var2"), value.name = "Correlation", threshold = NULL, sort = FALSE)
}
\arguments{
\item{m}{A matrix or an object that can be coerced to one (like a
sparseMatrix)}

\item{use}{Passed on to \code{\link{cor}}}

\item{method}{Passed on to \code{\link{cor}}}

\item{varnames}{Names for variable columns, passed on to \code{reshape2::melt}}

\item{threshold}{Optionally, a minimum threshold to include}

\item{sort}{Whether to sort the results in descending correlation}

\item{varnames}{Name for value column, passed on to \code{reshape2::melt}}
}
\description{
Perform a correlation and melt the result into a table
}

