% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorized_tests.R
\name{vectorized_t_test}
\alias{vectorized_t_test}
\title{Note that this comes from functions I previously wrote here:
https://github.com/dgrtwo/GSEAMA/blob/master/R/TestAssociation.R
Perform a Student's T-test comparing a metric to each column of a matrix}
\usage{
vectorized_t_test(m, y, var.equal = FALSE, alternative = "two.sided")
}
\arguments{
\item{m}{binary matrix}

\item{y}{numeric vector to test each column of m against}

\item{var.equal}{assume that the variances are equal}

\item{alternative}{whether to use a one-sided test, and if so which way}
}
\description{
This function is similar to \code{apply(m, 2, function(col) t.test(y ~ col)$p.value)},
but is vectorized to make it much faster
}
