% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vectorized_tests.R
\name{vectorized_wilcoxon_test}
\alias{vectorized_wilcoxon_test}
\title{Perform a Wilcoxon rank-sum test comparing a metric to each column of a matrix}
\usage{
vectorized_wilcoxon_test(m, y, alternative = "two.sided", tbl = FALSE)
}
\arguments{
\item{m}{binary matrix}

\item{y}{numeric vector to test each column of m against}

\item{alternative}{whether to use a one-sided test, and if so which way}

\item{tbl}{whether to return a data.frame rather than a vector}
}
\description{
This function is similar to \code{apply(m, 2, function(col) wilcox.test(y ~ col)$p.value)},
but is vectorized to make it much faster
}

