% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-strata.R
\name{extract_strata}
\alias{extract_strata}
\title{Extract strata from a tidied survival object}
\usage{
extract_strata(tbl, column, ...)
}
\arguments{
\item{tbl}{A table}

\item{column}{Column to extract. Should be of the form \code{col1=val1, col2=val2, ...}}

\item{...}{Extra arguments passed on to \link[tidyr]{extract}, such as \code{remove} or
\code{convert}.}
}
\description{
The tidied results of a survival object are in
}
\examples{

library(survival)
library(broom)

surv_td <- survfit(Surv(stop - start, status) ~ sex + pcdx,
                    data = mgus1) \%>\%
  tidy()

surv_td

# Extract into "sex" and "pcdx" columns
surv_fit \%>\%
  extract_strata(strata)

# Now visualize, separating by strata
library(ggplot2)
surv_fit \%>\%
  extract_strata(strata) \%>\%
  ggplot(aes(time, estimate, color = sex)) +
  geom_line() +
  geom_ribbon(aes(ymin = conf.low, ymax = conf.high), alpha = .25) +
  facet_wrap(~ pcdx)

}
